Updated version - Now patches exe file
It's all online now. It downloads Steamless and NTCore 4GB Patch, patches the exe file to remove DRM and add 4GBRam support, and also does the MTFix downloading the needed files from this github. As it finishes, it deletes itself, leaving only the new needed files, a modified exe file and an unmodified exe as "grid.bak"